+++ src/scvi/model/_condscvi.py
@@ -5,7 +5,6 @@ import warnings
 from typing import TYPE_CHECKING
 
 import numpy as np
-import pandas as pd
 import torch
 
 from scvi import REGISTRY_KEYS, settings
@@ -15,23 +14,22 @@ from scvi.data.fields import CategoricalObsField, LabelsWithUnlabeledObsField, L
 from scvi.model.base import (
     BaseModelClass,
     RNASeqMixin,
-    UnsupervisedTrainingMixin,
+    SemisupervisedTrainingMixin,
     VAEMixin,
 )
 from scvi.module import VAEC
 from scvi.utils import setup_anndata_dsp
-from scvi.utils._docstrings import devices_dsp
 
 if TYPE_CHECKING:
-    from collections.abc import Sequence
-
     from anndata import AnnData
 
 logger = logging.getLogger(__name__)
 
 
-class CondSCVI(RNASeqMixin, VAEMixin, UnsupervisedTrainingMixin, BaseModelClass):
+class CondSCVI(RNASeqMixin, VAEMixin, SemisupervisedTrainingMixin, BaseModelClass):
     """Conditional version of single-cell Variational Inference.
+
+    Used for multi-resolution deconvolution of spatial transcriptomics data :cite:p:`Lopez22`.
 
     Parameters
     ----------
@@ -64,13 +62,9 @@ class CondSCVI(RNASeqMixin, VAEMixin, UnsupervisedTrainingMixin, BaseModelClass)
     See further usage examples in the following tutorial:
 
     1. :doc:`/tutorials/notebooks/spatial/DestVI_tutorial`
-
     """
 
     _module_cls = VAEC
-    _LATENT_QZM = "_condscvi_latent_qzm"
-    _LATENT_QZV = "_condscvi_latent_qzv"
-    _OBSERVED_LIB_SIZE = "_condscvi_observed_lib_size"
 
     def __init__(
         self,
@@ -84,14 +78,9 @@ class CondSCVI(RNASeqMixin, VAEMixin, UnsupervisedTrainingMixin, BaseModelClass)
     ):
         super().__init__(adata)
 
-        n_batch = self.summary_stats.n_batch
-        n_labels = self.summary_stats.n_labels
-        n_vars = self.summary_stats.n_vars
-        if "n_fine_labels" in self.summary_stats:
-            self.n_fine_labels = self.summary_stats.n_fine_labels
-        else:
-            self.n_fine_labels = None
-        self._set_indices_and_labels(adata)
+        self.n_labels = self.summary_stats.n_labels
+        self.n_vars = self.summary_stats.n_vars
+        self._set_indices_and_labels()
         if weight_obs:
             ct_counts = np.unique(
                 self.get_from_registry(adata, REGISTRY_KEYS.LABELS_KEY),
@@ -104,10 +93,10 @@ class CondSCVI(RNASeqMixin, VAEMixin, UnsupervisedTrainingMixin, BaseModelClass)
             module_kwargs.update({"ct_weight": ct_weight})
 
         self.module = self._module_cls(
-            n_input=n_vars,
-            n_batch=n_batch,
-            n_labels=n_labels,
-            n_fine_labels=self.n_fine_labels,
+            n_input=self.n_vars,
+            n_batch=getattr(self.summary_stats, "n_batch", 0),
+            n_labels=self.n_labels,
+            n_fine_labels=getattr(self.summary_stats, "n_fine_labels", None),
             n_hidden=n_hidden,
             n_latent=n_latent,
             n_layers=n_layers,
@@ -227,203 +216,206 @@ class CondSCVI(RNASeqMixin, VAEMixin, UnsupervisedTrainingMixin, BaseModelClass)
 
         return results
 
-    @torch.inference_mode()
-    def predict(
-        self,
-        adata: AnnData | None = None,
-        indices: list[int] | None = None,
-        soft: bool = False,
-        batch_size: int | None = None,
-        use_posterior_mean: bool = True,
-    ) -> np.ndarray | pd.DataFrame:
-        """Return cell label predictions.
+    # @torch.inference_mode()
+    # def predict(
+    #     self,
+    #     adata: AnnData | None = None,
+    #     indices: list[int] | None = None,
+    #     soft: bool = False,
+    #     batch_size: int | None = None,
+    #     use_posterior_mean: bool = True,
+    # ) -> np.ndarray | pd.DataFrame:
+    #     """Return cell label predictions.
+    #
+    #     Parameters
+    #     ----------
+    #     adata
+    #         AnnData object that has been registered via :meth:`~scvi.model.SCANVI.setup_anndata`.
+    #     indices
+    #         Return probabilities for each class label.
+    #     soft
+    #         If True, returns per class probabilities
+    #     batch_size
+    #         Minibatch size for data loading into model. Defaults to `scvi.settings.batch_size`.
+    #     use_posterior_mean
+    #         If ``True``, uses the mean of the posterior distribution to predict celltype
+    #         labels. Otherwise, uses a sample from the posterior distribution - this
+    #         means that the predictions will be stochastic.
+    #     """
+    #
+    #     adata = self._validate_anndata(adata)
+    #
+    #     if indices is None:
+    #         indices = np.arange(adata.n_obs)
+    #
+    #     scdl = self._make_data_loader(
+    #         adata=adata,
+    #         indices=indices,
+    #         batch_size=batch_size,
+    #     )
+    #     y_pred = []
+    #     for _, tensors in enumerate(scdl):
+    #         inference_input = self.module._get_inference_input(tensors)
+    #         qz = self.module.inference(**inference_input)["qz"]
+    #         if use_posterior_mean:
+    #             z = qz.loc
+    #         else:
+    #             z = qz.sample()
+    #         pred = self.module.classify(
+    #             z,
+    #             label_index=inference_input["y"],
+    #         )
+    #         if self.module.classifier.logits:
+    #             pred = torch.nn.functional.softmax(pred, dim=-1)
+    #         if not soft:
+    #             pred = pred.argmax(dim=1)
+    #         y_pred.append(pred.detach().cpu())
+    #
+    #     y_pred = torch.cat(y_pred).numpy()
+    #     if not soft:
+    #         predictions = []
+    #         for p in y_pred:
+    #             predictions.append(self._code_to_fine_label[p])
+    #
+    #         return np.array(predictions)
+    #     else:
+    #         n_labels = len(pred[0])
+    #         pred = pd.DataFrame(
+    #             y_pred,
+    #             columns=self._fine_label_mapping[:n_labels],
+    #             index=adata.obs_names[indices],
+    #         )
+    #         return pred
 
-        Parameters
-        ----------
-        adata
-            AnnData object that has been registered via :meth:`~scvi.model.SCANVI.setup_anndata`.
-        indices
-            Return probabilities for each class label.
-        soft
-            If True, returns per class probabilities
-        batch_size
-            Minibatch size for data loading into model. Defaults to `scvi.settings.batch_size`.
-        use_posterior_mean
-            If ``True``, uses the mean of the posterior distribution to predict celltype
-            labels. Otherwise, uses a sample from the posterior distribution - this
-            means that the predictions will be stochastic.
-        """
-        adata = self._validate_anndata(adata)
+    # @torch.inference_mode()
+    # def confusion_coarse_celltypes(
+    #     self,
+    #     adata: AnnData | None = None,
+    #     indices: Sequence[int] | None = None,
+    #     batch_size: int | None = None,
+    # ) -> np.ndarray | pd.DataFrame:
+    #     """Return likelihood ratios to assess if switch coarse cell-type resolution
+    #
+    #     is too granular
+    #
+    #     Parameters
+    #     ----------
+    #     adata
+    #         AnnData object that has been registered via :meth:`~scvi.model.SCANVI.setup_anndata`.
+    #     indices
+    #         Return probabilities for each class label.
+    #     soft
+    #         If True, returns per class probabilities
+    #     batch_size
+    #         Minibatch size for data loading into model. Defaults to `scvi.settings.batch_size`.
+    #     use_posterior_mean
+    #         If ``True``, uses the mean of the posterior distribution to predict celltype
+    #         labels. Otherwise, uses a sample from the posterior distribution - this
+    #         means that the predictions will be stochastic.
+    #
+    #     """
+    #     adata = self._validate_anndata(adata)
+    #
+    #     if indices is None:
+    #         indices = np.arange(adata.n_obs)
+    #
+    #     scdl = self._make_data_loader(
+    #         adata=adata,
+    #         indices=indices,
+    #         batch_size=batch_size,
+    #     )
+    #     # Iterate once over the data and computes the reconstruction error
+    #     keys = list(self._label_mapping) + ["original"]
+    #     log_lkl = {key: [] for key in keys}
+    #     for tensors in scdl:
+    #         loss_kwargs = {"kl_weight": 1}
+    #         _, _, losses = self.module(tensors, loss_kwargs=loss_kwargs)
+    #         log_lkl["original"] += [losses.reconstruction_loss]
+    #         for i in range(self.module.n_labels):
+    #             tensors_ = tensors
+    #             tensors_["y"] = torch.full_like(tensors["y"], i)
+    #             _, _, losses = self.module(tensors_, loss_kwargs=loss_kwargs)
+    #             log_lkl[keys[i]] += [losses.reconstruction_loss]
+    #     for key in keys:
+    #         log_lkl[key] = torch.stack(log_lkl[key]).detach().numpy()
+    #
+    #     return log_lkl
 
-        if indices is None:
-            indices = np.arange(adata.n_obs)
+    # @devices_dsp.dedent
+    # def train(
+    #     self,
+    #     max_epochs: int = 300,
+    #     lr: float = 0.001,
+    #     accelerator: str = "auto",
+    #     devices: int | list[int] | str = "auto",
+    #     train_size: float = 1,
+    #     validation_size: float | None = None,
+    #     shuffle_set_split: bool = True,
+    #     batch_size: int = 128,
+    #     datasplitter_kwargs: dict | None = None,
+    #     plan_kwargs: dict | None = None,
+    #     **kwargs,
+    # ):
+    #     """Trains the model using MAP inference.
+    #
+    #     Parameters
+    #     ----------
+    #     max_epochs
+    #         Number of epochs to train for
+    #     lr
+    #         Learning rate for optimization.
+    #     %(param_accelerator)s
+    #     %(param_devices)s
+    #     train_size
+    #         Size of training set in the range [0.0, 1.0].
+    #     validation_size
+    #         Size of the test set. If `None`, defaults to 1 - `train_size`. If
+    #         `train_size + validation_size < 1`, the remaining cells belong to a test set.
+    #     shuffle_set_split
+    #         Whether to shuffle indices before splitting. If `False`, the val, train, and test set
+    #         are split in the sequential order of the data according to `validation_size` and
+    #         `train_size` percentages.
+    #     batch_size
+    #         Minibatch size to use during training.
+    #     datasplitter_kwargs
+    #         Additional keyword arguments passed into :class:`~scvi.dataloaders.DataSplitter`.
+    #     plan_kwargs
+    #         Keyword args for :class:`~scvi.train.TrainingPlan`. Keyword arguments passed to
+    #         `train()` will overwrite values present in `plan_kwargs`, when appropriate.
+    #     **kwargs
+    #         Other keyword args for :class:`~scvi.train.Trainer`.
+    #     """
+    #     update_dict = {
+    #         "lr": lr,
+    #     }
+    #     if plan_kwargs is not None:
+    #         plan_kwargs.update(update_dict)
+    #     else:
+    #         plan_kwargs = update_dict
+    #     super().train(
+    #         max_epochs=max_epochs,
+    #         accelerator=accelerator,
+    #         devices=devices,
+    #         train_size=train_size,
+    #         validation_size=validation_size,
+    #         shuffle_set_split=shuffle_set_split,
+    #         batch_size=batch_size,
+    #         datasplitter_kwargs=datasplitter_kwargs,
+    #         plan_kwargs=plan_kwargs,
+    #         **kwargs,
+    #     )
 
-        scdl = self._make_data_loader(
-            adata=adata,
-            indices=indices,
-            batch_size=batch_size,
-        )
-        y_pred = []
-        for _, tensors in enumerate(scdl):
-            inference_input = self.module._get_inference_input(tensors)
-            qz = self.module.inference(**inference_input)["qz"]
-            if use_posterior_mean:
-                z = qz.loc
-            else:
-                z = qz.sample()
-            pred = self.module.classify(
-                z,
-                label_index=inference_input["y"],
-            )
-            if self.module.classifier.logits:
-                pred = torch.nn.functional.softmax(pred, dim=-1)
-            if not soft:
-                pred = pred.argmax(dim=1)
-            y_pred.append(pred.detach().cpu())
-
-        y_pred = torch.cat(y_pred).numpy()
-        if not soft:
-            predictions = []
-            for p in y_pred:
-                predictions.append(self._code_to_fine_label[p])
-
-            return np.array(predictions)
-        else:
-            n_labels = len(pred[0])
-            pred = pd.DataFrame(
-                y_pred,
-                columns=self._fine_label_mapping[:n_labels],
-                index=adata.obs_names[indices],
-            )
-            return pred
-
-    @torch.inference_mode()
-    def confusion_coarse_celltypes(
-        self,
-        adata: AnnData | None = None,
-        indices: Sequence[int] | None = None,
-        batch_size: int | None = None,
-    ) -> np.ndarray | pd.DataFrame:
-        """Return likelihood ratios to assess if switch coarse cell-type resolution is too granular
-
-        Parameters
-        ----------
-        adata
-            AnnData object that has been registered via :meth:`~scvi.model.SCANVI.setup_anndata`.
-        indices
-            Return probabilities for each class label.
-        soft
-            If True, returns per class probabilities
-        batch_size
-            Minibatch size for data loading into model. Defaults to `scvi.settings.batch_size`.
-        use_posterior_mean
-            If ``True``, uses the mean of the posterior distribution to predict celltype
-            labels. Otherwise, uses a sample from the posterior distribution - this
-            means that the predictions will be stochastic.
-
-        """
-        adata = self._validate_anndata(adata)
-
-        if indices is None:
-            indices = np.arange(adata.n_obs)
-
-        scdl = self._make_data_loader(
-            adata=adata,
-            indices=indices,
-            batch_size=batch_size,
-        )
-        # Iterate once over the data and computes the reconstruction error
-        keys = list(self._label_mapping) + ["original"]
-        log_lkl = {key: [] for key in keys}
-        for tensors in scdl:
-            loss_kwargs = {"kl_weight": 1}
-            _, _, losses = self.module(tensors, loss_kwargs=loss_kwargs)
-            log_lkl["original"] += [losses.reconstruction_loss]
-            for i in range(self.module.n_labels):
-                tensors_ = tensors
-                tensors_["y"] = torch.full_like(tensors["y"], i)
-                _, _, losses = self.module(tensors_, loss_kwargs=loss_kwargs)
-                log_lkl[keys[i]] += [losses.reconstruction_loss]
-        for key in keys:
-            log_lkl[key] = torch.stack(log_lkl[key]).detach().numpy()
-
-        return log_lkl
-
-    @devices_dsp.dedent
-    def train(
-        self,
-        max_epochs: int = 300,
-        lr: float = 0.001,
-        accelerator: str = "auto",
-        devices: int | list[int] | str = "auto",
-        train_size: float = 1,
-        validation_size: float | None = None,
-        shuffle_set_split: bool = True,
-        batch_size: int = 128,
-        datasplitter_kwargs: dict | None = None,
-        plan_kwargs: dict | None = None,
-        **kwargs,
-    ):
-        """Trains the model using MAP inference.
-
-        Parameters
-        ----------
-        max_epochs
-            Number of epochs to train for
-        lr
-            Learning rate for optimization.
-        %(param_accelerator)s
-        %(param_devices)s
-        train_size
-            Size of training set in the range [0.0, 1.0].
-        validation_size
-            Size of the test set. If `None`, defaults to 1 - `train_size`. If
-            `train_size + validation_size < 1`, the remaining cells belong to a test set.
-        shuffle_set_split
-            Whether to shuffle indices before splitting. If `False`, the val, train, and test set
-            are split in the sequential order of the data according to `validation_size` and
-            `train_size` percentages.
-        batch_size
-            Minibatch size to use during training.
-        datasplitter_kwargs
-            Additional keyword arguments passed into :class:`~scvi.dataloaders.DataSplitter`.
-        plan_kwargs
-            Keyword args for :class:`~scvi.train.TrainingPlan`. Keyword arguments passed to
-            `train()` will overwrite values present in `plan_kwargs`, when appropriate.
-        **kwargs
-            Other keyword args for :class:`~scvi.train.Trainer`.
-        """
-        update_dict = {
-            "lr": lr,
-        }
-        if plan_kwargs is not None:
-            plan_kwargs.update(update_dict)
-        else:
-            plan_kwargs = update_dict
-        super().train(
-            max_epochs=max_epochs,
-            accelerator=accelerator,
-            devices=devices,
-            train_size=train_size,
-            validation_size=validation_size,
-            shuffle_set_split=shuffle_set_split,
-            batch_size=batch_size,
-            datasplitter_kwargs=datasplitter_kwargs,
-            plan_kwargs=plan_kwargs,
-            **kwargs,
-        )
-
-    def _set_indices_and_labels(self, adata: AnnData):
-        """Set indices for labeled and unlabeled cells."""
-        labels_state_registry = self.adata_manager.get_state_registry(REGISTRY_KEYS.LABELS_KEY)
-        self.original_label_key = labels_state_registry.original_key
-        self._label_mapping = labels_state_registry.categorical_mapping
-        self._code_to_label = dict(enumerate(self._label_mapping))
-        if self.n_fine_labels is not None:
-            fine_labels_state_registry = self.adata_manager.get_state_registry("fine_labels")
-            self.original_fine_label_key = fine_labels_state_registry.original_key
-            self._fine_label_mapping = fine_labels_state_registry.categorical_mapping
-            self._code_to_fine_label = dict(enumerate(self._fine_label_mapping))
+    # def _set_indices_and_labels(self, adata: AnnData):
+    #     """Set indices for labeled and unlabeled cells."""
+    #     labels_state_registry = self.adata_manager.get_state_registry(REGISTRY_KEYS.LABELS_KEY)
+    #     self.original_label_key = labels_state_registry.original_key
+    #     self._label_mapping = labels_state_registry.categorical_mapping
+    #     self._code_to_label = dict(enumerate(self._label_mapping))
+    #     if self.n_fine_labels is not None:
+    #         fine_labels_state_registry = self.adata_manager.get_state_registry("fine_labels")
+    #         self.original_fine_label_key = fine_labels_state_registry.original_key
+    #         self._fine_label_mapping = fine_labels_state_registry.categorical_mapping
+    #         self._code_to_fine_label = dict(enumerate(self._fine_label_mapping))
 
     @classmethod
     @setup_anndata_dsp.dedent
@@ -446,9 +438,8 @@ class CondSCVI(RNASeqMixin, VAEMixin, UnsupervisedTrainingMixin, BaseModelClass)
         %(param_labels_key)s
         fine_labels_key
             Key in `adata.obs` where fine-grained labels are stored.
-        %(unlabeled_category)ss
         %(param_layer)s
-        %(param_batch_key)s
+        %(param_unlabeled_category)s
         """
         setup_method_args = cls._get_setup_method_args(**locals())
         anndata_fields = [