+++ CHANGELOG.md
@@ -24,7 +24,7 @@ to [Semantic Versioning]. Full commit history is available in the
 
 #### Changed
 
-- Update model {class}`scvi.model.DestVI` with fine cell-type classifier {pr}`33XX`.
+- Update model {class}`scvi.model.DestVI` with fine cell-type classifier {pr}`3380`.
 
 #### Removed